requirements.txt
pandas
numpy
pypdf
python-docx
flask
flask-cors
flask-restful
flask-sqlalchemy
python-dotenv
flask-login
flask-bcrypt
flask-wtf
flask-mail
django
mysqlclient
bs4
faiss-cpu
sentence-transformers
groq
hugchat
torch
streamlit

